added defer fn for login error auditing
